requirements.txt
numpy==2.2.*
gradio==5.15.*
pillow==11.1.*
transformers==4.50.*
torch==2.6.*
accelerate==1.5.*
ollama~=0.5.*